Break out portable common.SimpleServer
